+++ src/main/java/module-info.java
@@ -82,6 +82,7 @@
 module org.cicirello.chips_n_salsa {
 	exports org.cicirello.search;
 	exports org.cicirello.search.concurrent;
+	exports org.cicirello.search.evo;
 	exports org.cicirello.search.hc;
 	exports org.cicirello.search.operators;
 	exports org.cicirello.search.operators.bits;
@@ -101,4 +102,6 @@ module org.cicirello.chips_n_salsa {
 	// return Permutation objects, defined in this
 	// module.
 	requires transitive org.cicirello.jpt;
+	requires org.cicirello.rho_mu;
+	requires org.cicirello.core;
 }