+++ src/main/java/org/cicirello/search/hc/FirstDescentHillClimber.java
@@ -138,10 +138,11 @@ public final class FirstDescentHillClimber<T extends Copyable<T>> extends Abstra
 					if (!keepClimbing) iter.rollback();
 				}
 				// update tracker
+				boolean isMinCost = pOptInt.isMinCost(currentCost);
 				if (currentCost < tracker.getCost()) {
-					tracker.update(currentCost, current, pOptInt.isMinCost(currentCost));
+					tracker.update(currentCost, current, isMinCost);
 				}
-				return new SolutionCostPair<T>(current, currentCost);
+				return new SolutionCostPair<T>(current, currentCost, isMinCost);
 		};
 	}
 	
@@ -167,10 +168,11 @@ public final class FirstDescentHillClimber<T extends Copyable<T>> extends Abstra
 					if (!keepClimbing) iter.rollback();
 				}
 				// update tracker
+				boolean isMinCost = pOpt.isMinCost(currentCost);
 				if (currentCost < tracker.getCostDouble()) {
-					tracker.update(currentCost, current, pOpt.isMinCost(currentCost));
+					tracker.update(currentCost, current, isMinCost);
 				}			
-				return new SolutionCostPair<T>(current, currentCost);
+				return new SolutionCostPair<T>(current, currentCost, isMinCost);
 		};
 	}
 }