+++ src/test/java/org/cicirello/search/concurrent/ParallelMultistarterTwoThreadsTests.java
@@ -513,14 +513,14 @@ public class ParallelMultistarterTwoThreadsTests {
 		public SolutionCostPair<TestObject> optimize(int runLength) {
 			optCounter++;
 			int c = update(runLength);
-			return new SolutionCostPair<TestObject>(new TestObject(), c);
+			return new SolutionCostPair<TestObject>(new TestObject(), c, false);
 		}
 		
 		@Override
 		public SolutionCostPair<TestObject> reoptimize(int runLength) {	
 			reoptCounter++;
 			int c = update(runLength);
-			return new SolutionCostPair<TestObject>(new TestObject(), c);
+			return new SolutionCostPair<TestObject>(new TestObject(), c, false);
 		}
 		
 		private int update(int runLength) {